add srt_only & video_only

+++ src/cli/main.py
@@ -16,33 +16,6 @@ def main():
 
 @main.command()
 @click.option("--url", "-u", required=True, help="Youtube URL.")
-@click.option(
-    "--threads",
-    "-n",
-    default=8,
-    help="Number of threads to use",
-)
-@click.option(
-    "--format",
-    "-f",
-    default="bestvideo+bestaudio/best",
-    help="Download format. See: https://github.com/yt-dlp/yt-dlp#format-selection-examples",
-)
-def video(url: str, threads: int, format: str) -> None:
-    """Download video from Youtube."""
-    download(url, threads, format)
-
-
-@main.command()
-@click.option("--url", "-u", required=True, help="Youtube URL.")
-@click.option("--threads", "-n", default=8, help="Number of threads to use.")
-def audio(url: str, threads: int) -> None:
-    """Download video from Youtube."""
-    download_audio(url, threads)
-
-
-@main.command()
-@click.option("--url", "-u", required=True, help="Youtube URL.")
 @click.option("--threads", "-n", default=8, help="Number of threads to use.")
 @click.option(
     "--format",
@@ -63,10 +36,30 @@ def audio(url: str, threads: int) -> None:
     default=None,
     help="Root directory for the models.",
 )
+@click.option(
+    "--srt_only",
+    "-s",
+    default=False,
+    help="Only generate subtitles. Do not download video.",
+)
+@click.option(
+    "--video_only",
+    "-v",
+    default=False,
+    help="Only download video. Do not generate subtitles.",
+)
 def full(
-    url: str, threads: int, format: str, model_name: str, model_root: str | None
+    url: str,
+    threads: int,
+    format: str,
+    model_name: str,
+    model_root: str | None,
+    srt_only: bool,
+    video_only: bool,
 ) -> None:
     """Download video from Youtube and generate subtitles."""
+
+    # Check if URL is valid
     try:
         yt_dlp.YoutubeDL().extract_info(
             url, download=False
@@ -75,26 +68,39 @@ def full(
         click.echo("Invalid URL. Please provide a valid YouTube URL.")
         return
 
-    logger.info(f"Downloading video from {url}.")
-    audio_file_name = download_audio(url, threads)
-    srt_file_name = audio_file_name.split(".")[0] + ".srt"
-
     # Download video in parallel
-    download_thread = Thread(target=download, args=(url, threads, format))
-    download_thread.start()
+    download_thread = None
+    if not srt_only:
+        # Download video in parallel
+        download_thread = Thread(target=download, args=(url, threads, format))
+        download_thread.start()
+    else:
+        logger.info("Skipping video download due to --srt_only flag.")
 
     # Transcribe audio
-    if not os.path.exists(srt_file_name):
-        logger.info(f"Transcribing audio from {audio_file_name}.")
-        model = initialize_whisper_model(model_name=model_name, model_root=model_root)
-        segs, info = model.transcribe(audio=audio_file_name, vad_filter=True)
-        write_srt(segments=segs, file=srt_file_name)
-        if os.path.exists(audio_file_name):
-            os.remove(audio_file_name)
-    else:
-        logger.info(f"File {srt_file_name} already exists. Skipping transcription.")
+    logger.info(f"Downloading video from {url}.")
 
-    download_thread.join()
+    if not video_only:
+        audio_file_name = download_audio(url, threads)
+        srt_file_name = audio_file_name.split(".")[0] + ".srt"
+        if not os.path.exists(srt_file_name):
+            logger.info(f"Transcribing audio from {audio_file_name}.")
+            model = initialize_whisper_model(
+                model_name=model_name, model_root=model_root
+            )
+            segs, info = model.transcribe(audio=audio_file_name, vad_filter=True)
+            write_srt(segments=segs, file=srt_file_name)
+            if os.path.exists(audio_file_name):
+                os.remove(audio_file_name)
+        else:
+            logger.info(
+                f"Skipping audio transcription due to {srt_file_name} already exists."
+            )
+    else:
+        logger.info("Skipping audio transcription due to --video_only flag.")
+
+    if download_thread is not None:
+        download_thread.join()
 
 
 if __name__ == "__main__":
